refactor(basic): split semantic analyzer implementation
